refactor(e2e): move provisioner funcs to provisioner.go, writeCheckFile to helpers
